fix: Refactor Audit to CodeAudit
