+++ src/main/java/acme/entities/codeAudit/Mark.java
@@ -1,5 +1,5 @@
 
-package acme.entities.audit;
+package acme.entities.codeAudit;
 
 public enum Mark {
 	APlus, A, B, C, F, FMinus;
