+++ src/main/java/acme/entities/codeAudit/Type.java
@@ -1,5 +1,5 @@
 
-package acme.entities.audit;
+package acme.entities.codeAudit;
 
 public enum Type {
 	STATIC, DYNAMIC;
